move process candidate starter model classes to process model impl module
